Solved test failures related to latest changes in the security relates services
