+++ ripple-security/src/test/java/org/rippleosi/login/service/LoginServiceTest.java
@@ -1,7 +1,5 @@
 package org.rippleosi.login.service;
 
-import java.net.URI;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -11,14 +9,11 @@ import org.pac4j.core.context.J2EContext;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.profile.ProfileManager;
 import org.pac4j.core.profile.UserProfile;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 public class LoginServiceTest {
@@ -106,34 +101,5 @@ public class LoginServiceTest {
         final List<String> permissions = userProfile.getPermissions();
 
         assertNotNull("The user's permissions have not been saved in the user profile.", permissions);
-    }
-
-    @Test
-    public void shouldNotGenerateLocationHeaderIfRedirectUrlIsMalformed() {
-        final String invalidUrl = "Malformed" + "\n" + "URL";
-
-        final ResponseEntity<String> responseEntity =
-            loginService.generateRedirectResponseEntity(invalidUrl, null, HttpStatus.I_AM_A_TEAPOT);
-
-        assertNull("The redirect location header must not be set if invalid.", responseEntity.getHeaders().getLocation());
-    }
-
-    @Test
-    public void shouldExpandRedirectUrlIfParamsIncluded() {
-        final String expectedUrl = "http://www.google.co.uk/search?param1=test&param2=test";
-
-        final String inputUrlBasePath = "http://www.google.co.uk/search";
-
-        final Map<String, String> params = new HashMap<>();
-        params.put("param1", "test");
-        params.put("param2", "test");
-
-        final ResponseEntity<String> responseEntity =
-            loginService.generateRedirectResponseEntity(inputUrlBasePath, params, HttpStatus.I_AM_A_TEAPOT);
-
-        final URI outputUrl = responseEntity.getHeaders().getLocation();
-
-        assertNotNull("The redirect location URL must not be null using valid inputs.", outputUrl);
-        assertEquals("The redirect location URL must be fully expanded using the given params.", expectedUrl, outputUrl.toString());
     }
 }
